Refactor Modal from JS to TS
